+++ setup_sklearnex.py
@@ -81,7 +81,7 @@ setup(name="scikit-learn-intelex",
           'sklearnex.cluster',
           'sklearnex.decomposition',
           'sklearnex.ensemble',
-          'sklearnex.global',
+          'sklearnex.globally',
           'sklearnex.linear_model',
           'sklearnex.manifold',
           'sklearnex.metrics',
