+++ sklearnex/dispatcher.py
@@ -52,13 +52,13 @@ def get_patch_names():
     return list(get_patch_map().keys())
 
 
-def patch_sklearn(name=None, verbose=True, _global=False):
+def patch_sklearn(name=None, verbose=True, global_patch=False):
     if not sklearn_check_version('0.22'):
         raise NotImplementedError("Intel(R) Extension for Scikit-learn* patches apply "
                                   "for scikit-learn >= 0.22 only ...")
 
-    if _global:
-        from ._utils import patch_sklearn_global
+    if global_patch:
+        from sklearnex.globally.dispatcher import patch_sklearn_global
         patch_sklearn_global(name, verbose)
         return
 
@@ -77,9 +77,9 @@ def patch_sklearn(name=None, verbose=True, _global=False):
             "(https://github.com/intel/scikit-learn-intelex)\n")
 
 
-def unpatch_sklearn(name=None, _global=False):
-    if _global:
-        from ._utils import unpatch_sklearn_global
+def unpatch_sklearn(name=None, global_patch=False):
+    if global_patch:
+        from sklearnex.globally.dispatcher import unpatch_sklearn_global
         unpatch_sklearn_global()
     from daal4py.sklearn import unpatch_sklearn as unpatch_sklearn_orig
     if isinstance(name, list):
